Supports #29 effort to break up common_stuff into two modules, the new one is show_text
