+++ lib/printsly.rb
@@ -1,10 +1,12 @@
 require 'choice'
 require 'colorize'
 require 'common_stuff'
+require 'show_text'
 require 'json'
 require 'menu'
 require 'printers'
 require 'spreadsheet'
 Spreadsheet.client_encoding = 'UTF-8'
 
+include ShowText
 include CommonStuff
